--- js/tempCodeRunnerFile.js
@@ -1,2 +1,0 @@
-console.log("Available island groups:", Object.keys(locationData));
-
